Split up roads.mss, refactor road colour script
The generated colour values now reside in road-colors-generated.mss, and
have been removed from roads.mms. road-colors-generated.mms is generated
by the colour script, and should not be edited manually. To regenerate
it, edit road-colors.yaml and run:

```
./scripts/generate_road_colours.py > road-colors-generated.mss
```

The original output of the colour script provided some info about the
amount of loss caused by converting from CIE lch to RGB. To get this
output, call the script with the `-v` flag (for verbose).